requirements.txt
influxdb-client[ciso]
python-miio @ git+https://github.com/rytilahti/python-miio.git
python-dotenv
Adafruit-DHT
getmac
gpiozero